feat: Added New DB Functions
