+++ server/database/user.go
@@ -7,6 +7,29 @@ import (
 	"github.com/surrealdb/surrealdb.go"
 )
 
+func (db *DB) AddNewUser(fullName string, email string, hashedPassword string) (*models.User, error) {
+
+	params := map[string]interface{}{
+		"fullName": fullName,
+		"email":    email,
+		"password": hashedPassword,
+	}
+
+	rawData, err := db.client.Create("user", params)
+	if err != nil {
+		return nil, fmt.Errorf("unable to create User in database: %v", err)
+	}
+
+	var user []models.User
+
+	err = surrealdb.Unmarshal(rawData, &user)
+	if err != nil {
+		return nil, fmt.Errorf("unable to parse data from database: %v", err)
+	}
+
+	return &user[0], nil
+}
+
 func (db *DB) GetUserByEmail(email string) (*models.User, error) {
 
 	query := "SELECT * FROM user WHERE email = $email"
@@ -45,38 +68,13 @@ func (db *DB) GetProfileByID(id string) (*models.Profile, error) {
 		return nil, fmt.Errorf("unable to fetch User: %v", err)
 	}
 
-	var user []models.Profile
+	var user models.Profile
 
 	err = surrealdb.Unmarshal(rawData, &user)
 	if err != nil {
 		return nil, fmt.Errorf("unable to unmarshal: %v", err)
 	}
 
-	if len(user) == 0 {
-		return nil, fmt.Errorf("user not found")
-	}
-
-	return &user[0], nil
+	return &user, nil
 
 }
-
-// func (db *DB) GetUserByID(id string) (*models.User, error) {
-// 	rawData, err := db.client.Select(id)
-// 	if err != nil {
-// 		return nil, fmt.Errorf("unable to fetch User: %v", err)
-// 	}
-
-// 	var user []models.Profile
-
-// 	err = surrealdb.Unmarshal(rawData, &user)
-// 	if err != nil {
-// 		return nil, fmt.Errorf("unable to unmarshal: %v", err)
-// 	}
-
-// 	if len(user) == 0 {
-// 		return nil, fmt.Errorf("user not found")
-// 	}
-
-// 	return &user[0], nil
-
-// }
